// frontend/src/shared/api/generated/models/PrepareTicketRefundRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger TON Lottery
 * TON Lottery
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: almostinf@mail.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PrepareTicketRefundRequest
 */
export interface PrepareTicketRefundRequest {
    /**
     * 
     * @type {string}
     * @memberof PrepareTicketRefundRequest
     */
    idempotencyKey: string;
    /**
     * 
     * @type {string}
     * @memberof PrepareTicketRefundRequest
     */
    walletId: string;
    /**
     * 
     * @type {string}
     * @memberof PrepareTicketRefundRequest
     */
    ticketId: string;
}

/**
 * Check if a given object implements the PrepareTicketRefundRequest interface.
 */
export function instanceOfPrepareTicketRefundRequest(value: object): value is PrepareTicketRefundRequest {
    if (!('idempotencyKey' in value) || value['idempotencyKey'] === undefined) return false;
    if (!('walletId' in value) || value['walletId'] === undefined) return false;
    if (!('ticketId' in value) || value['ticketId'] === undefined) return false;
    return true;
}

export function PrepareTicketRefundRequestFromJSON(json: any): PrepareTicketRefundRequest {
    return PrepareTicketRefundRequestFromJSONTyped(json, false);
}

export function PrepareTicketRefundRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PrepareTicketRefundRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'idempotencyKey': json['idempotencyKey'],
        'walletId': json['walletId'],
        'ticketId': json['ticketId'],
    };
}

export function PrepareTicketRefundRequestToJSON(json: any): PrepareTicketRefundRequest {
    return PrepareTicketRefundRequestToJSONTyped(json, false);
}

export function PrepareTicketRefundRequestToJSONTyped(value?: PrepareTicketRefundRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'idempotencyKey': value['idempotencyKey'],
        'walletId': value['walletId'],
        'ticketId': value['ticketId'],
    };
}