// frontend/src/shared/api/generated/models/PrepareTicketRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger TON Lottery
 * TON Lottery
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: almostinf@mail.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PrepareTicketRequest
 */
export interface PrepareTicketRequest {
    /**
     * 
     * @type {string}
     * @memberof PrepareTicketRequest
     */
    idempotencyKey: string;
    /**
     * 
     * @type {string}
     * @memberof PrepareTicketRequest
     */
    walletId: string;
    /**
     * 
     * @type {string}
     * @memberof PrepareTicketRequest
     */
    lotteryId: string;
}

/**
 * Check if a given object implements the PrepareTicketRequest interface.
 */
export function instanceOfPrepareTicketRequest(value: object): value is PrepareTicketRequest {
    if (!('idempotencyKey' in value) || value['idempotencyKey'] === undefined) return false;
    if (!('walletId' in value) || value['walletId'] === undefined) return false;
    if (!('lotteryId' in value) || value['lotteryId'] === undefined) return false;
    return true;
}

export function PrepareTicketRequestFromJSON(json: any): PrepareTicketRequest {
    return PrepareTicketRequestFromJSONTyped(json, false);
}

export function PrepareTicketRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PrepareTicketRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'idempotencyKey': json['idempotencyKey'],
        'walletId': json['walletId'],
        'lotteryId': json['lotteryId'],
    };
}

export function PrepareTicketRequestToJSON(json: any): PrepareTicketRequest {
    return PrepareTicketRequestToJSONTyped(json, false);
}

export function PrepareTicketRequestToJSONTyped(value?: PrepareTicketRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'idempotencyKey': value['idempotencyKey'],
        'walletId': value['walletId'],
        'lotteryId': value['lotteryId'],
    };
}