// frontend/src/shared/api/generated/models/PrepareTicketTransactionResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger TON Lottery
 * TON Lottery
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: almostinf@mail.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PrepareTicketTransactionResponseTx } from './PrepareTicketTransactionResponseTx';
import {
    PrepareTicketTransactionResponseTxFromJSON,
    PrepareTicketTransactionResponseTxFromJSONTyped,
    PrepareTicketTransactionResponseTxToJSON,
    PrepareTicketTransactionResponseTxToJSONTyped,
} from './PrepareTicketTransactionResponseTx';

/**
 * 
 * @export
 * @interface PrepareTicketTransactionResponse
 */
export interface PrepareTicketTransactionResponse {
    /**
     * 
     * @type {string}
     * @memberof PrepareTicketTransactionResponse
     */
    intentId: string;
    /**
     * 
     * @type {string}
     * @memberof PrepareTicketTransactionResponse
     */
    ticketId: string;
    /**
     * 
     * @type {string}
     * @memberof PrepareTicketTransactionResponse
     */
    lotteryId: string;
    /**
     * 
     * @type {Date}
     * @memberof PrepareTicketTransactionResponse
     */
    expiresAt: Date;
    /**
     * 
     * @type {PrepareTicketTransactionResponseTx}
     * @memberof PrepareTicketTransactionResponse
     */
    tx: PrepareTicketTransactionResponseTx;
}

/**
 * Check if a given object implements the PrepareTicketTransactionResponse interface.
 */
export function instanceOfPrepareTicketTransactionResponse(value: object): value is PrepareTicketTransactionResponse {
    if (!('intentId' in value) || value['intentId'] === undefined) return false;
    if (!('ticketId' in value) || value['ticketId'] === undefined) return false;
    if (!('lotteryId' in value) || value['lotteryId'] === undefined) return false;
    if (!('expiresAt' in value) || value['expiresAt'] === undefined) return false;
    if (!('tx' in value) || value['tx'] === undefined) return false;
    return true;
}

export function PrepareTicketTransactionResponseFromJSON(json: any): PrepareTicketTransactionResponse {
    return PrepareTicketTransactionResponseFromJSONTyped(json, false);
}

export function PrepareTicketTransactionResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PrepareTicketTransactionResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'intentId': json['intentId'],
        'ticketId': json['ticketId'],
        'lotteryId': json['lotteryId'],
        'expiresAt': (new Date(json['expiresAt'])),
        'tx': PrepareTicketTransactionResponseTxFromJSON(json['tx']),
    };
}

export function PrepareTicketTransactionResponseToJSON(json: any): PrepareTicketTransactionResponse {
    return PrepareTicketTransactionResponseToJSONTyped(json, false);
}

export function PrepareTicketTransactionResponseToJSONTyped(value?: PrepareTicketTransactionResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'intentId': value['intentId'],
        'ticketId': value['ticketId'],
        'lotteryId': value['lotteryId'],
        'expiresAt': value['expiresAt'].toISOString(),
        'tx': PrepareTicketTransactionResponseTxToJSON(value['tx']),
    };
}